+++ src/Console/RePushNotificationCommand.php
@@ -42,10 +42,9 @@ class RePushNotificationCommand extends Command
                     'notification_id' => $notification->id,
                 ];
 
-                $response = FcmNotification::sendNotification($notification, $message);
+                FcmNotification::sendNotification($notification, $message);
 
                 $notification->update([
-                    'response' => $response,
                     'status' => 'completed'
                 ]);
